Cambios formato modulo baja usuarios y cambios modulo clientes. Eliminacion form de mas (restablecer contraseña)
